Fix app warning user (incorrectly) that boot completed intent was not received
Some manufacturer's implement aggressive battery optimizations that hinder apps starting in background.
The deactivated alert dialog warned the user if that was the case and suggested to disable these optimization - to allow p!n to show the users note's as notifications directly after boot.
Due to a bug in the app this warning dialog did sometimes also appear even if the app was started on boot (probably because it was at some point stopped by the system and reloaded when the user reopened it)

Disable this warning dialog completely as a (temporary) fix.

Upgrade version in manifest, build.gradle and github release tag to 1.0.1 because the app is right now being submitted to fdroid so that the fixed, updated version is distributed should the app be accepted.

Besides this now removed "feature" - the warning dialog - the app is thoroughly tested and *should* run stable.

Changes to be committed:
	modified:   app/build.gradle
	modified:   app/src/main/AndroidManifest.xml
	modified:   app/src/main/java/de/nproth/pin/NoteActivity.java