Fix pin snooze duration not updating
Problem caused by the singleton Pinboard not being a singleton.
In the last bugfix release Pinboard had to be accessed directly because p!n is not allowed to start the encapsulating PinboardService in background in ContentReceivers on newer android versions.
Thus we ended up with multiple unsynchronized versions of Pinboard.
This fix writes changed values back to the SharedPreferences.
But Pinboard is still not synchronizing between its instances.